% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toyData.R
\docType{data}
\name{toyData}
\alias{toyData}
\title{Simulated dyadic data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4950 rows and 13 columns.
}
\usage{
data(toyData)
}
\description{
Simulated data with 100 units and full pairwise dyads.
}
\keyword{datasets}
