% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastDyadRobust.R
\name{vcovDyadRobust}
\alias{vcovDyadRobust}
\title{Compute dyadic-robust standard errors}
\usage{
vcovDyadRobust(
  fit,
  dyad_cluster,
  nthreads = RcppParallel::defaultNumThreads() - 1
)
}
\arguments{
\item{fit}{The model object estimated by the \code{fixest} package.}

\item{dyad_cluster}{A data frame or matrix that specifies how to dyadically
cluster the standard errors.
The dimension of \code{dyad_cluster} must be D x 2,
where D is the number of rows of the data frame that you have used
to estimate \code{fit}.}

\item{nthreads}{The number of threads. The default is to use all threads minus one.}
}
\description{
Computes dyad-robust standard errors via multiway decomposition,
proposed by Aronow, Peter M., Cyrus Samii, and Valentina A. Assenova.
"Cluster-robust variance estimation for dyadic data."
Political Analysis 23.4 (2015): 564-577.

This package is based on \code{dyadRobust}.
}
