% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovDyadRobust.R
\name{vcovDyadRobust}
\alias{vcovDyadRobust}
\title{Compute dyadic-robust standard errors}
\usage{
vcovDyadRobust(
  fit,
  dyad_cluster,
  nthreads = RcppParallel::defaultNumThreads() - 1
)
}
\arguments{
\item{fit}{The model object. Can be of any class compatible with the \code{sandwich} package.}

\item{dyad_cluster}{A data frame or matrix that specifies how to dyadically
cluster the standard errors.
The dimension of \code{dyad_cluster} must be D x 2,
where D is the number of rows of the data frame that you have used
to estimate \code{fit}.}

\item{nthreads}{The number of threads. The default is to use all threads minus one.}
}
\value{
Returns dyadic-robust standard errors calculated via multiway decomposition.
}
\description{
Computes dyadic-robust standard errors via multiway decomposition
proposed by Aronow, Samii, and Assenova (2015).

This function is based on \code{dyadRobust()} from the \code{dyadRobust} package.
See \url{https://github.com/jbisbee1/dyadRobust} for details.
}
\details{
For details of dyadic-robust standard errors see the original paper:

Aronow, Peter M., Cyrus Samii, and Valentina A. Assenova.
"Cluster-robust variance estimation for dyadic data."
Political Analysis 23.4 (2015): 564-577.
}
\examples{
\dontrun{
# Load a toy dataset
df <- fastDyadRobust::toyData

# Run regression
reg <- lm(dy ~ dx1 + dx2, df)

# Prepare data containing node ids
df_dyad <- df[c("src", "dst")]

# Get dyadic-robust standard errors
fastDyadRobust::vcovDyadRobust(reg, df_dyad)
}

}
